ml/demand-forecasting/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
xgboost==2.0.3
statsmodels==0.14.2
prophet==1.1.5
tensorflow==2.17.0
boto3==1.34.100
confluent-kafka==2.4.0
httpx==0.27.0
python-dotenv==1.0.1
structlog==24.2.0
pytest==8.2.0
pytest-cov==5.0.0
pytest-asyncio==0.23.7
ruff==0.4.4
slowapi==0.1.9
PyJWT==2.8.0
bandit==1.7.9
safety==3.2.3
